Remove test keys after running tests
