+++ utils/security/keychain_test.go
@@ -19,15 +19,13 @@ import (
 	"github.com/zalando/go-keyring"
 )
 
-const testDeployment = "LOCAL"
-const testUsername = "unit_test_user"
 const testPassword = "pAss%-w0rd-&'cha*s"
 const testPasswordUpdated = "pAss%-w0rd-&'cha*s-with_more_chars"
 
 func Test_Keychain(t *testing.T) {
 
 	// remove test key if one exists
-	keyring.Delete(KeyringServiceName+"."+testDeployment, testUsername)
+	keyring.Delete(strings.ToLower(KeyringServiceName+"."+testDeployment), testUsername)
 
 	t.Run("Secret can not be retrieved for an unknown account", func(t *testing.T) {
 		retrievedSecret, err := SecKeyGetSecret(testDeployment, testUsername)
